+++ src/main/java/com/scc/dto/Title.java
@@ -1,0 +1,36 @@
+package com.scc.dto;
+
+import org.eclipse.microprofile.openapi.annotations.media.Schema;
+
+import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
+
+@JsonIgnoreProperties(ignoreUnknown = true)
+public class Title extends Hidden {
+
+    @Schema(required = true, description = "code")
+    private String code;
+    @Schema(required = true, description = "name")
+    private String name;
+
+    public String getCode() {
+        return code;
+    }
+
+    public void setCode(String code) {
+        this.code = code;
+    }
+
+    public String getName() {
+        return name;
+    }
+
+    public void setName(String name) {
+        this.name = name;
+    }
+
+    @Override
+    public String toString() {
+        return "Title [code=" + code + ", name=" + name + ", hidden=" + hidden + "]";
+    }
+
+}
