+++ src/main/java/com/scc/health/LivenessResource.java
@@ -1,0 +1,18 @@
+package com.scc.health;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.health.HealthCheck;
+import org.eclipse.microprofile.health.HealthCheckResponse;
+import org.eclipse.microprofile.health.Liveness;
+
+@ApplicationScoped
+@Liveness
+public class LivenessResource implements HealthCheck {
+
+    @Override
+    public HealthCheckResponse call() {
+        return HealthCheckResponse.named("WS is alive !").up().build();
+    }
+
+}
