+++ src/main/java/com/scc/health/ReadinessResource.java
@@ -1,0 +1,18 @@
+package com.scc.health;
+
+import javax.enterprise.context.ApplicationScoped;
+
+import org.eclipse.microprofile.health.HealthCheck;
+import org.eclipse.microprofile.health.HealthCheckResponse;
+import org.eclipse.microprofile.health.Readiness;
+
+@ApplicationScoped
+@Readiness
+public class ReadinessResource implements HealthCheck {
+
+    @Override
+    public HealthCheckResponse call() {
+        return HealthCheckResponse.named("WS is ready !").up().build();
+    }
+
+}
