+++ src/main/java/com/scc/rest/PedigreeResource.java
@@ -1,4 +1,4 @@
-package com.scc.controller;
+package com.scc.rest;
 
 import javax.annotation.security.RolesAllowed;
 import javax.enterprise.context.ApplicationScoped;
@@ -23,8 +23,8 @@ import org.eclipse.microprofile.openapi.annotations.parameters.Parameter;
 import org.jboss.logging.Logger;
 import org.jboss.resteasy.annotations.cache.NoCache;
 
+import com.scc.dto.Dog;
 import com.scc.exception.NotFoundException;
-import com.scc.model.Dog;
 import com.scc.service.DogService;
 
 @Path("/api")
@@ -32,9 +32,9 @@ import com.scc.service.DogService;
 @ApplicationScoped
 @Consumes(MediaType.APPLICATION_JSON)
 @Produces(MediaType.APPLICATION_JSON)
-public class PedigreeController {
+public class PedigreeResource {
 
-    private static final Logger LOG = Logger.getLogger(PedigreeController.class);
+    private static final Logger LOG = Logger.getLogger(PedigreeResource.class);
 
     @Inject
     DogService dogService;
