+++ src/main/resources/application.properties
@@ -20,12 +20,20 @@ quarkus.native.report-exception-stack-traces=true
 %prod.quarkus.oidc.client-id=ws-pedigree-app
 %prod.quarkus.oidc.credentials.secret=cebc914a-e582-44f4-af6c-61bcf8b9eab5
 %prod.quarkus.http.cors=true
+
 # Enable Policy Enforcement
 %prod.quarkus.keycloak.policy-enforcer.enable=true
+
+quarkus.keycloak.policy-enforcer.enforcement-mode=PERMISSIVE
 
 # Disables policy enforcement for a path
 quarkus.keycloak.policy-enforcer.paths.1.path=/public/*
 quarkus.keycloak.policy-enforcer.paths.1.enforcement-mode=DISABLED
+
+#quarkus.keycloak.policy-enforcer.paths.health.name=Heath
+#quarkus.keycloak.policy-enforcer.paths.health.path=/health/*
+#quarkus.keycloak.policy-enforcer.paths.health.enforcement-mode=DISABLED
+
 
 # Database
 quarkus.hibernate-orm.database.generation=drop-and-create
