+++ src/test/java/com/scc/keycloak/PolicyEnforcerTest.java
@@ -1,0 +1,65 @@
+package com.scc.keycloak;
+
+import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.extension.ExtendWith;
+import org.keycloak.representations.AccessTokenResponse;
+
+import io.quarkus.test.junit.QuarkusTest;
+import io.restassured.RestAssured;
+
+@ExtendWith(KeycloakServer.class)
+@QuarkusTest
+public class PolicyEnforcerTest {
+
+    private static final String KEYCLOAK_SERVER_URL = System.getProperty("keycloak.url", "http://localhost:8180/auth");
+    private static final String KEYCLOAK_REALM = "quarkus";
+
+    @Test
+    public void testAccessUserResource() {
+        RestAssured.given().auth().oauth2(getAccessToken("alice"))
+                .when().get("/api/users/me")
+                .then()
+                .statusCode(200);
+        RestAssured.given().auth().oauth2(getAccessToken("jdoe"))
+                .when().get("/api/users/me")
+                .then()
+                .statusCode(200);
+    }
+
+    @Test
+    public void testAccessAdminResource() {
+        RestAssured.given().auth().oauth2(getAccessToken("alice"))
+                .when().get("/api/admin")
+                .then()
+                .statusCode(403);
+        RestAssured.given().auth().oauth2(getAccessToken("jdoe"))
+                .when().get("/api/admin")
+                .then()
+                .statusCode(403);
+        RestAssured.given().auth().oauth2(getAccessToken("admin"))
+                .when().get("/api/admin")
+                .then()
+                .statusCode(200);
+    }
+
+    @Test
+    public void testPublicResource() {
+        RestAssured.given()
+                .when().get("/api/public")
+                .then()
+                .statusCode(204);
+    }
+
+    private String getAccessToken(String userName) {
+        return RestAssured
+                .given()
+                .param("grant_type", "password")
+                .param("username", userName)
+                .param("password", userName)
+                .param("client_id", "backend-service")
+                .param("client_secret", "secret")
+                .when()
+                .post(KEYCLOAK_SERVER_URL + "/realms/" + KEYCLOAK_REALM + "/protocol/openid-connect/token")
+                .as(AccessTokenResponse.class).getToken();
+    }
+}
