+++ src/test/resources/import.sql
@@ -1,0 +1,7 @@
+INSERT INTO pg_arbre_genealogie (id_chien,id_pere,id_mere) VALUES (1725763, 1725476, 1725231);
+
+INSERT INTO pg_data (iddog,data) VALUES (1725763, '{"id" : 1725763, "name" : "ELGA", "gender" : "F", "birthDate" : "1989-03-28", "birthCountry" : "FR", "inbreedingCoefficient" : "0%", "pedigrees" : [{   "country" : "FR",   "type" : "LOF",   "number" : "98993/17939" }], "tokens" : [{   "type" : "tatoo",   "number" : "JVP685" }], "breed" : {   "idRace" : 223,   "codeFci" : "002" }, "health" : [{   "code" : "HD-A",   "libelle" : "" }], "breeder" : {   "civility" : "M.",   "lastName" : "LONDEIX" }, "titles" : [{   "code" : "CHA",   "name" : "CHAMPION DES FIELD-TRIALS D''AUTOMNE" }, {   "code" : "CHCS",   "name" : "CHAMPION DE FRANCE DE CONFORMITE AU STANDARD" }, {   "code" : "TR",   "name" : "TRIALER" }]}')
+INSERT INTO pg_data (iddog,data) VALUES (1725476, '{"id" : 1725476, "name" : "ALMER LOI. 286872", "gender" : "M", "birthDate" : "1984-04-03", "birthCountry" : "FR", "inbreedingCoefficient" : "0%", "pedigrees" : [{   "country" : "FR",   "type" : "LOF",   "number" : "81739/9431" }], "tokens" : [{   "type" : "tatoo",   "number" : "EFE479" }], "breed" : {   "idRace" : 223,   "codeFci" : "002" }, "health" : [{   "code" : "HD",   "libelle" : "" }], "titles" : [{   "code" : "CHQFS",   "name" : "CHAMPION QUETE A LA FRANCAISE SOLO" }, {   "code" : "CHQFC",   "name" : "CHAMPION QUETE FRANCAISE COUPLE" }, {   "code" : "CHCPC",   "name" : "CHAMPION CHASSE PRATIQUE COUPLE" }, {   "code" : "TR",   "name" : "TRIALER" }]}')
+INSERT INTO pg_data (iddog,data) VALUES (1725231, '{"id" : 1725231, "name" : "AGATE", "gender" : "F", "birthDate" : "1985-08-13", "birthCountry" : "FR", "inbreedingCoefficient" : "0%", "pedigrees" : [{   "country" : "FR",   "type" : "LOF",   "number" : "84238/13321" }], "tokens" : [{   "type" : "tatoo",   "number" : "FVE378" }], "breed" : {   "idRace" : 223,   "codeFci" : "002" }, "health" : [{   "code" : "HD",   "libelle" : "" }], "breeder" : {   "civility" : "M.",   "lastName" : "LONDEIX" }, "titles" : [{   "code" : "TR",   "name" : "TRIALER" }]}')
+
+
